+++ dio/CHANGELOG.md
@@ -10,6 +10,7 @@ See the [Migration Guide][] for the complete breaking changes list.**
 - Create type alias `DioMediaType` for `http_parser`'s `MediaType`.
 - Fix the type conversion regression when using `MultipartFile.fromBytes`.
 - Add FusedTransformer for improved performance when decoding JSON.
+- Improves `InterceptorState.toString()`.
 
 ## 5.4.3+1